Remove Video option from Multimedia Post; adjust cropped image view
- from Discord

"The issue is with retrieving the URI from given file path.

At Edit Profile side, Glide seems to handle this on our behalf. So what is being used on Edit Profile can be reused for the Multimedia Post.

To do the Video, a custom method to retrieve URI (aka pinpoint the exact file to upload from the phone's directory) has to be created. I have been able to find methods which can get the URI of videos stored on Google Drive, but not those stored on an external SD card & internal storage of the phone."